requirements.txt
python-dotenv
mcp-client
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.18
pydantic==2.5.0
openai==1.3.7